+++ src/redux/reviewsSlice.js
@@ -5,22 +5,28 @@ import axios from 'axios';
 // O BACKEND REAL (Node + MongoDB)
 const API_URL = 'http://localhost:8000/reviews';
 
+const authConfig = (getState) => {
+  const token = getState().auth?.token;
+  return {
+    headers: {
+      Authorization: token ? `Bearer ${token}` : '',
+    },
+  };
+};
+
 // Criar review COM JWT
 export const createReview = createAsyncThunk(
     'reviews/createReview',
     async (reviewData, { rejectWithValue }) => {
         try {
+            const config = authConfig(getState);
             const token = localStorage.getItem('token');
 
             if (!token) {
                 return rejectWithValue('Usuário não autenticado.');
             }
 
-            const response = await axios.post(API_URL, reviewData, {
-                headers: {
-                    Authorization: `Bearer ${token}`,
-                },
-            });
+            const response = await axios.post(API_URL, reviewData, config);
 
             return response.data;
         } catch (err) {
@@ -34,13 +40,10 @@ export const fetchMyReviews = createAsyncThunk(
     'reviews/fetchMyReviews',
     async (_, { rejectWithValue }) => {
         try {
+            const config = authConfig(getState);
             const token = localStorage.getItem('token');
 
-            const response = await axios.get(`${API_URL}/me`, {
-                headers: {
-                    Authorization: `Bearer ${token}`,
-                },
-            });
+            const response = await axios.get(`${API_URL}/me`, config);
 
             return response.data;
         } catch (err) {
